Update project.py
A simple project using python to calculate the sgpa of a student and  store all the details in students_sgpa.xlsx  file for further access of those data. At first we have to give the name, registration number , branch , and after that grade of all subjects.
Then after submitting the input it will showing the sgpa in pop up box . Also available clear button to clear all after one entry.
And these details are stored in xlsx file in a formal way.